requirements.txt
flask
gunicorn
requests
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
rapidfuzz